requirements.txt
Appium-Python-Client==0.28
cpf-generator==1.0.0
pytest==3.8.2
selenium==3.14.1
six==1.11.0
